Tests/LibWeb/Text/expected/css/style-invalidation/sheet-add-font-face-only.txt

@font-face only sheet add: fullStyleInvalidations=0, styleInvalidations=3, elementStyleRecomputations=2, elementStyleNoopRecomputations=1
